HZN-1163: Restructure modules so that the adapter api no longer depends on protobuf (since adapters themselves may depend on different versions of protobuf).
