+++ features/telemetry/config/api/pom.xml
@@ -2,13 +2,13 @@
 <project xmlns="http://maven.apache.org/POM/4.0.0" xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance" xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 http://maven.apache.org/maven-v4_0_0.xsd">
   <parent>
     <groupId>org.opennms.features.telemetry</groupId>
-    <artifactId>org.opennms.features.telemetry.adapters</artifactId>
-    <version>22.0.0-SNAPSHOT</version>
+    <artifactId>org.opennms.features.telemetry.config</artifactId>
+    <version>21.0.0-SNAPSHOT</version>
   </parent>
   <modelVersion>4.0.0</modelVersion>
-  <groupId>org.opennms.features.telemetry.adapters</groupId>
-  <artifactId>org.opennms.features.telemetry.adapters.api</artifactId>
-  <name>OpenNMS :: Features :: Telemetry :: Adapters :: API</name>
+  <groupId>org.opennms.features.telemetry.config</groupId>
+  <artifactId>org.opennms.features.telemetry.config.api</artifactId>
+  <name>OpenNMS :: Features :: Telemetry :: Config :: API</name>
   <packaging>bundle</packaging>
   <build>
     <plugins>
@@ -26,11 +26,4 @@
       </plugin>
     </plugins>
   </build>
-  <dependencies>
-    <dependency>
-      <groupId>org.opennms.features.telemetry.config</groupId>
-      <artifactId>org.opennms.features.telemetry.config.api</artifactId>
-      <version>${project.version}</version>
-    </dependency>
-  </dependencies>
 </project>
