+++ features/telemetry/config/api/src/main/java/org/opennms/netmgt/telemetry/config/api/Package.java
@@ -26,55 +26,26 @@
  *     http://www.opennms.com/
  *******************************************************************************/
 
-package org.opennms.netmgt.telemetry.ipc;
-
-import java.util.Optional;
+package org.opennms.netmgt.telemetry.config.api;
 
 /**
- * Protocol configuration details that are required
- * by both OpenNMS and Minion.
+ * Telemetry protocol package configuration.
  */
-public interface ProtocolDefinition {
+public interface Package {
 
     /**
-     * The name of the protocol.
+     * The filter rule is used to match which sources should belong to this package.
      *
-     * This is used as a suffix for any associated queues that are created and
-     * must be the same on both OpenNMS and Minion.
+     * If the rule is <code>null</code>, then all sources should match.
      *
-     * @return the protocol name
+     * @return the fitler rule
      */
-    String getName();
+    String getFilterRule();
 
     /**
-     * Number of threads used for consuming/dispatching messages.
+     * The RRD settings are use to control the control of RRD files, when applicable.
      *
-     * @return the number of threads
+     * @return the rrd settings
      */
-    Optional<Integer> getNumThreads();
-
-    /**
-     * Messages are aggregated in batches before being dispatched.
-     * When the batch reaches this size, it will be dispatched.
-     *
-     * @return the batch size
-     */
-    Optional<Integer> getBatchSize();
-
-    /**
-     * Messages are aggregated in batches before being dispatched.
-     * When the batch has been created for longer than this interval (ms)
-     * it will be dispatched, regardless of the current size.
-     *
-     * @return the batch interval
-     */
-    Optional<Integer> getBatchIntervalMs();
-
-    /**
-     * Maximum number of messages to keep in memory while waiting
-     * to be dispatched.
-     *
-     * @return the queue size
-     */
-    Optional<Integer> getQueueSize();
+    Rrd getRrd();
 }
