+++ features/telemetry/config/jaxb/src/assembly/xsds.xml
