+++ features/telemetry/config/jaxb/src/main/java/org/opennms/netmgt/telemetry/config/model/Adapter.java
