+++ features/telemetry/config/jaxb/src/main/resources/org/opennms/netmgt/telemetry/config/model/jaxb.properties
