+++ features/telemetry/config/jaxb/src/main/resources/xsds/telemetryd-config.xsd
