+++ features/telemetry/config/pom.xml
@@ -8,90 +8,10 @@
   <modelVersion>4.0.0</modelVersion>
   <groupId>org.opennms.features.telemetry</groupId>
   <artifactId>org.opennms.features.telemetry.config</artifactId>
-  <name>OpenNMS :: Features :: Telemetry :: Config</name>
-  <packaging>bundle</packaging>
-  <build>
-    <plugins>
-      <plugin>
-        <groupId>org.apache.felix</groupId>
-        <artifactId>maven-bundle-plugin</artifactId>
-        <extensions>true</extensions>
-        <configuration>
-          <instructions>
-            <Bundle-RequiredExecutionEnvironment>JavaSE-1.8</Bundle-RequiredExecutionEnvironment>
-            <Bundle-SymbolicName>${project.artifactId}</Bundle-SymbolicName>
-            <Bundle-Version>${project.version}</Bundle-Version>
-          </instructions>
-        </configuration>
-      </plugin>
-<!--  No need to generate xsds everytime, use pre-compiled one
-      <plugin>
-        <groupId>org.codehaus.mojo</groupId>
-        <artifactId>jaxb2-maven-plugin</artifactId>
-        <version>1.6</version>
-        <executions>
-          <execution>
-            <id>schemagen</id>
-            <goals>
-              <goal>schemagen</goal>
-            </goals>
-          </execution>
-        </executions>
-        <configuration>
-          <transformSchemas>
-            <transformSchema>
-              <uri>http://xmlns.opennms.org/xsd/config/telemetryd-config</uri>
-              <toFile>telemetryd-config.xsd</toFile>
-            </transformSchema>
-          </transformSchemas>
-          <outputDirectory>${project.build.directory}/xsds</outputDirectory>
-          <workDirectory>${project.build.directory}/jaxb2-work</workDirectory>
-          <includes>
-            <include>org/opennms/netmgt/telemetry/config/model/*.java</include>
-          </includes>
-          <verbose>true</verbose>
-        </configuration>
-      </plugin> -->
-      <plugin>
-        <artifactId>maven-assembly-plugin</artifactId>
-        <extensions>true</extensions>
-        <configuration>
-          <descriptors>
-            <descriptor>src/assembly/xsds.xml</descriptor>
-          </descriptors>
-          <tarLongFileMode>posix</tarLongFileMode>
-        </configuration>
-        <executions>
-          <execution>
-            <phase>package</phase>
-            <goals>
-              <goal>single</goal>
-            </goals>
-          </execution>
-        </executions>
-	  </plugin>
-    </plugins>
-  </build>
-  <dependencies>
-    <dependency>
-      <groupId>org.opennms.core</groupId>
-      <artifactId>org.opennms.core.xml</artifactId>
-    </dependency>
-    <dependency>
-      <groupId>org.opennms.features.telemetry.listeners</groupId>
-      <artifactId>org.opennms.features.telemetry.listeners.api</artifactId>
-      <version>${project.version}</version>
-    </dependency>
-    <dependency>
-      <groupId>org.opennms.features.telemetry</groupId>
-      <artifactId>org.opennms.features.telemetry.common</artifactId>
-      <version>${project.version}</version>
-    </dependency>
-
-    <dependency>
-      <groupId>org.opennms.core.test-api</groupId>
-      <artifactId>org.opennms.core.test-api.xml</artifactId>
-      <scope>test</scope>
-    </dependency>
-  </dependencies>
+  <packaging>pom</packaging>
+  <name>OpenNMS :: Features :: Telemetry :: API</name>
+  <modules>
+    <module>api</module>
+    <module>jaxb</module>
+  </modules>
 </project>
