+++ features/telemetry/minion/pom.xml
@@ -33,6 +33,11 @@
       <version>${project.version}</version>
     </dependency>
     <dependency>
+      <groupId>org.opennms.features.telemetry.config</groupId>
+      <artifactId>org.opennms.features.telemetry.config.api</artifactId>
+      <version>${project.version}</version>
+    </dependency>
+    <dependency>
       <groupId>org.opennms.features.telemetry.listeners</groupId>
       <artifactId>org.opennms.features.telemetry.listeners.udp</artifactId>
       <version>${project.version}</version>
