+++ opennms-assemblies/xsds/src/assembly/xsds.xml
@@ -48,7 +48,7 @@
         <include>org.opennms.protocols:org.opennms.protocols.nsclient:jar:xsds:${project.version}</include>
         <include>org.opennms.protocols:org.opennms.protocols.xml:jar:xsds:${project.version}</include>
         <include>org.opennms.protocols:org.opennms.protocols.xmp:jar:xsds:${project.version}</include>
-        <include>org.opennms.features.telemetry:org.opennms.features.telemetry.config:jar:xsds:${project.version}</include>
+        <include>org.opennms.features.telemetry.config:org.opennms.features.telemetry.config.jaxb:jar:xsds:${project.version}</include>
       </includes>
     </dependencySet>
   </dependencySets>
