+++ opennms-config-tester/pom.xml
@@ -104,8 +104,8 @@
       <artifactId>opennms-alarm-northbounder-snmptrap</artifactId>
     </dependency>
     <dependency>
-      <groupId>org.opennms.features.telemetry</groupId>
-      <artifactId>org.opennms.features.telemetry.config</artifactId>
+      <groupId>org.opennms.features.telemetry.config</groupId>
+      <artifactId>org.opennms.features.telemetry.config.jaxb</artifactId>
       <version>${project.version}</version>
     </dependency>
     <dependency>
